Conflicts:
	tensorflow/compiler/xla/python/shared_device_buffer.cc
	tensorflow/compiler/xla/service/gpu/ir_emitter.cc
	tensorflow/compiler/xla/service/gpu/ir_emitter_unnested.cc
	tensorflow/compiler/xla/service/gpu/tests/BUILD
	tensorflow/python/distribute/collective_all_reduce_strategy_test.py
	tensorflow/python/keras/layers/BUILD
	tensorflow/python/keras/layers/recurrent_v2.py
	tensorflow/python/kernel_tests/signal/fft_ops_test.py

It looks like you may be committing a merge.
If this is not correct, please remove the file
	.git/MERGE_HEAD
and try again.

+++ tensorflow/lite/micro/memory_planner/memory_planner.h
@@ -58,7 +58,7 @@ class MemoryPlanner {
                                  int last_time_used) = 0;
 
   // The largest contguous block of memory that's needed to hold the layout.
-  virtual int GetMaximumMemorySize() = 0;
+  virtual size_t GetMaximumMemorySize() = 0;
   // How many buffers have been added to the planner.
   virtual int GetBufferCount() = 0;
   // Calculated layout offset for the N-th buffer added to the planner.
